chore: Rename folders
