+++ src/helpers/encrypt-password.ts
