+++ src/models/mail-options.interface.ts
