+++ src/routes/register-user.ts
@@ -1,4 +1,4 @@
-import { CreateUser } from "controller";
+import { CreateUser } from "controllers";
 import { Router } from "express";
 
 const router = Router();
