+++ src/validators/user-equal-password.ts
@@ -1,4 +1,4 @@
-import { IUser } from "../model";
+import { IUser } from "../models";
 
 class EqualPasswordValidator
 {
